Refactor wallet related command Moved to separate file
